+++ pytest.ini
@@ -1,0 +1,4 @@
+[pytest]
+markers =
+    datadownload: test that downloads data from external sources
+    slow: mark test as slow
